+++ src/store/util.ts
@@ -18,3 +18,22 @@ export function autoMapStateToProps (state: any, names: string[]) {
     return obj;
   }, {});
 }
+
+
+
+export function buildReducer (initialState: any, actions: any) {
+  return (state: any = initialState, action: any) => {
+    const { type } = action;
+    if (actions[type]) {
+      const rv = Object.assign({}, state);
+      actions[type](rv, action);
+      return rv;
+    }
+    else if (type.startsWith('@@redux/')) {
+      return state;
+    }
+    else {
+      throw new Error(`Unknown action type: ${type}`);
+    }
+  };
+}
